Fixed `Player::HasScotopia`.
